+++ xzinternals/example.go
